+++ src/main/kotlin/codegen/FlutterXPluginInfo.kt
@@ -1,12 +1,12 @@
 package codegen
 
 object FlutterXPluginInfo {
-    const val Version: String = "5.3.2.242"
+    const val Version: String = "5.3.3.243"
     const val Changelog: String = """
-<h2>5.3.2 - 2025-03-04</h2>
+<h2>5.3.3 - 2025-03-10</h2>
 
-<ul><li>New: <code>flutter android gradle</code> Migration Tools (android tool)</li><li>Fix: <code>dio</code> The problem of listening server not being destroyed (dio tool)</li></ul>
+<ul><li>New: add assets preview window</li><li>fix: The problem of jitter in asset images exceeding a certain size (document preview)</li><li>fix: assets code generate tool setting errors</li><li>Some other problems solved</li></ul>
 """
 
-    const val BuildDate: String = "2025-03-04"
+    const val BuildDate: String = "2025-03-13"
 }