+++ src/main/kotlin/shop/itbug/fluttercheckversionx/config/PluginConfig.kt
@@ -1,56 +1,45 @@
 package shop.itbug.fluttercheckversionx.config
 
-import com.intellij.ide.plugins.RepositoryHelper
 import com.intellij.openapi.components.*
 import com.intellij.openapi.project.Project
-import shop.itbug.fluttercheckversionx.constance.ideaPluginStoreUrl
+import com.intellij.openapi.project.guessProjectDir
 
-data class PluginSetting(
-
-    //是否显示操作工具
-    var showRiverpodInlay: Boolean = true,
-
-    //自动导包的定义
-    var autoImportRiverpodText: String = "package:hooks_riverpod/hooks_riverpod.dart",
-
-    //在Android studio中打开android目录
-    var openAndroidDirectoryInAS: Boolean = true,
-    //在xcode中打开ios目录
-    var openIosDirectoryInXcode: Boolean = true,
-    //在xcode中打开macos目录
-    var openMacosDirectoryInXcode: Boolean = true,
-
-    //开启本地资产预览
-    var showAssetsIconInEditor: Boolean = true,
-    //缩放大小
-    var assetsIconSize: Int = 16,
-
-    //显示discord
-    var showDiscord: Boolean = true,
-    //显示qq群
-    var showQQGroup: Boolean = true,
-    //显示打赏action
-    var showRewardAction: Boolean = true,
-
-    //是否加入了开发版本更新
-    var isJoinDevVersion: Boolean = RepositoryHelper.getCustomPluginRepositoryHosts().contains(ideaPluginStoreUrl),
-
-    //是否启用 freezed3快速修复面板
-    var showFreezed3FixNotification: Boolean = true,
-) : BaseState()
+class PluginSetting : BaseState() {
+    var showRiverpodInlay by property(true)
+    var autoImportRiverpodText by string("package:hooks_riverpod/hooks_riverpod.dart")
+    var openAndroidDirectoryInAS by property(true)
+    var openIosDirectoryInXcode by property(true)
+    var openMacosDirectoryInXcode by property(true)
+    var showAssetsIconInEditor by property(true)
+    var assetsIconSize by property(16)
+    var showDiscord by property(true)
+    var showQQGroup by property(true)
+    var showRewardAction by property(true)
+    var showFreezed3FixNotification by property(true)
+    var enableAssetsPreviewAction by property(true)
+    var assetsPreviewImageSize by property(120)
+    var assetDirectory by string()
+}
 
 @Service(Service.Level.PROJECT)
 @State(name = "FlutterxFullConfig", storages = [Storage("FlutterxFullConfig.xml")])
-class PluginConfig : SimplePersistentStateComponent<PluginSetting>(PluginSetting()) {
-    private var setting = PluginSetting()
+class PluginConfig(val project: Project) : SimplePersistentStateComponent<PluginSetting>(PluginSetting()) {
 
-    override fun loadState(state: PluginSetting) {
-        setting = state
+
+    fun initAssetsDirectory() {
+        if (state.assetDirectory.isNullOrBlank()) {
+            state.assetDirectory =
+                (project.guessProjectDir()?.findChild("assets")?.path) ?: project.guessProjectDir()?.path
+            loadState(state)
+        }
     }
+
 
     companion object {
         fun getInstance(project: Project): PluginConfig {
-            return project.service<PluginConfig>()
+            val config = project.service<PluginConfig>()
+            config.initAssetsDirectory()
+            return config
         }
 
         fun getState(project: Project) = getInstance(project).state