+++ src/main/kotlin/shop/itbug/fluttercheckversionx/constance/StringConstance.kt
@@ -33,6 +33,11 @@ object Links {
 
     const val dioImage = "https://flutterx.itbug.shop/%E6%8E%A5%E5%8F%A3%E4%BF%A1%E6%81%AF%E6%88%AA%E5%9B%BE.html"
 
+    //资产预览窗口
+    const val assetsPreviewDoc =
+        "https://flutterx.itbug.shop/%E8%B5%84%E4%BA%A7%E9%A2%84%E8%A7%88%E7%AA%97%E5%8F%A3.html"
+
+
     //新版本检查
     const val checkFlutterVersion =
         "https://flutterx.itbug.shop/flutter%E6%96%B0%E7%89%88%E6%9C%AC%E6%A3%80%E6%B5%8B.html"