+++ src/main/kotlin/shop/itbug/fluttercheckversionx/dialog/JsonToDartBaseConfig.kt
@@ -85,6 +85,7 @@ fun Row.saveToDirectoryConfig(
     }
 }
 
+// freeze 新的设置
 object FreezedNewSetting {
     fun setting(project: Project, config: FreezedClassConfig, panel: Panel) {
         val box = ComboBox<FreezedVersion>(FreezedVersion.entries.toTypedArray())
@@ -103,27 +104,10 @@ object FreezedNewSetting {
 
 object MyRowBuild {
     /// 2024.3
-//    fun folder(row: Row, getter: () -> String, setter: (String) -> Unit, project: Project) {
-//        row.textFieldWithBrowseButton(
-//            FileChooserDescriptorFactory.createSingleFolderDescriptor(),
-//            project,
-//        ).bindText(getter, setter).align(Align.FILL).addValidationRule(VerifyFileDir.ERROR_MSG) {
-//            VerifyFileDir.validDirByComponent(it)
-//        }.validationOnInput {
-//            if (VerifyFileDir.validDirByComponent(it)) {
-//                return@validationOnInput ValidationInfoBuilder(it.textField).error(VerifyFileDir.ERROR_MSG)
-//            }
-//            return@validationOnInput null
-//        }
-//    }
-
-    /// 2023.2
     fun folder(row: Row, getter: () -> String, setter: (String) -> Unit, project: Project) {
         row.textFieldWithBrowseButton(
-            PluginBundle.get("select_a_folder"),
-            project,
             FileChooserDescriptorFactory.createSingleFolderDescriptor(),
-            { it.path },
+            project,
         ).bindText(getter, setter).align(Align.FILL).addValidationRule(VerifyFileDir.ERROR_MSG) {
             VerifyFileDir.validDirByComponent(it)
         }.validationOnInput {
@@ -133,6 +117,23 @@ object MyRowBuild {
             return@validationOnInput null
         }
     }
+
+    /// 2023.2
+//    fun folder(row: Row, getter: () -> String, setter: (String) -> Unit, project: Project) {
+//        row.textFieldWithBrowseButton(
+//            PluginBundle.get("select_a_folder"),
+//            project,
+//            FileChooserDescriptorFactory.createSingleFolderDescriptor(),
+//            { it.path },
+//        ).bindText(getter, setter).align(Align.FILL).addValidationRule(VerifyFileDir.ERROR_MSG) {
+//            VerifyFileDir.validDirByComponent(it)
+//        }.validationOnInput {
+//            if (VerifyFileDir.validDirByComponent(it)) {
+//                return@validationOnInput ValidationInfoBuilder(it.textField).error(VerifyFileDir.ERROR_MSG)
+//            }
+//            return@validationOnInput null
+//        }
+//    }
 }
 
 enum class NameFormatRule(val format: CaseFormat, val eg: String) {