+++ src/main/kotlin/shop/itbug/fluttercheckversionx/form/socket/MyDefaultListModel.kt
@@ -1,8 +1,6 @@
 package shop.itbug.fluttercheckversionx.form.socket
 
 import com.intellij.ui.ColoredListCellRenderer
-import com.intellij.ui.JBColor
-import com.intellij.ui.TextIcon
 import com.intellij.ui.components.JBLabel
 import shop.itbug.fluttercheckversionx.dsl.requestDetailLayout
 import java.awt.Component