+++ src/main/kotlin/shop/itbug/fluttercheckversionx/manager/FlutterAndroidMigrateManager.kt
@@ -13,7 +13,6 @@ import com.intellij.openapi.project.guessProjectDir
 import com.intellij.openapi.util.Key
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.openapi.vfs.readText
-import com.intellij.psi.PsiDocumentManager
 import com.intellij.psi.PsiManager
 import com.intellij.psi.util.PsiTreeUtil
 import com.intellij.testFramework.LightVirtualFile
@@ -62,7 +61,7 @@ class FlutterAndroidMigrateManager(val project: Project) : Disposable {
         WriteCommandAction.runWriteCommandAction(project) {
             val psiFile = PsiManager.getInstance(project).findFile(vf) as? GroovyFile ?: return@runWriteCommandAction
             psiFile.fileDocument.setText(newSettingFile.readText())
-            PsiDocumentManager.getInstance(project).commitDocument(psiFile.fileDocument)
+//            PsiDocumentManager.getInstance(project).commitDocument(psiFile.fileDocument)
         }
     }
 
@@ -148,10 +147,12 @@ class FlutterAndroidMigrateManager(val project: Project) : Disposable {
 }
 
 
+private val i = "includeBuild(\"\$flutterSdkPath/packages/flutter_tools/gradle\")"
+
 private val newSettingFile: VirtualFile
     get() {
         return LightVirtualFile(
-            "settings.gradle", FileTypeRegistry.getInstance().getFileTypeByExtension("gradle"), $$"""
+            "settings.gradle", FileTypeRegistry.getInstance().getFileTypeByExtension("gradle"), """
                 pluginManagement {
                     def flutterSdkPath = {
                         def properties = new Properties()
@@ -161,7 +162,7 @@ private val newSettingFile: VirtualFile
                         return flutterSdkPath
                     }()
 
-                    includeBuild("$flutterSdkPath/packages/flutter_tools/gradle")
+                    $i
 
                     repositories {
                         google()